@@ -1,21 +1,21 @@
 // AI CORE
 
 import * as fs from 'fs';
-import autobind from 'autobind-decorator';
-import * as loki from 'lokijs';
-import * as request from 'request-promise-native';
-import * as chalk from 'chalk';
+import { bindThis } from '@/decorators.js';
+import loki from 'lokijs';
+import got from 'got';
+import chalk from 'chalk';
 import { v4 as uuid } from 'uuid';
-const delay = require('timeout-as-promise');
 
-import config from '@/config';
-import Module from '@/module';
-import Message from '@/message';
-import Friend, { FriendDoc } from '@/friend';
-import { User } from '@/misskey/user';
-import Stream from '@/stream';
-import log from '@/utils/log';
-const pkg = require('../package.json');
+import config from '@/config.js';
+import Module from '@/module.js';
+import Message from '@/message.js';
+import Friend, { FriendDoc } from '@/friend.js';
+import { User } from '@/misskey/user.js';
+import Stream from '@/stream.js';
+import log from '@/utils/log.js';
+import { sleep } from './utils/sleep.js';
+import pkg from '../package.json' assert { type: 'json' };
 
 type MentionHook = (msg: Message) => Promise<boolean | HandlerResult>;
 type ContextHook = (key: any, msg: Message, data?: any) => Promise<void | boolean | HandlerResult>;
@@ -103,12 +103,12 @@ export default class 藍 {
 		});
 	}
 
-	@autobind
+	@bindThis
 	public log(msg: string) {
-		log(chalk`[{magenta AiOS}]: ${msg}`);
+		log(`[${chalk.magenta('AiOS')}]: ${msg}`);
 	}
 
-	@autobind
+	@bindThis
 	private run() {
 		//#region Init DB
 		this.meta = this.getCollection('meta', {});
@@ -207,7 +207,7 @@ export default class 藍 {
 	 * ユーザーから話しかけられたとき
 	 * (メンション、リプライ、トークのメッセージ)
 	 */
-	@autobind
+	@bindThis
 	private async onReceiveMessage(msg: Message): Promise<void> {
 		this.log(chalk.gray(`<<< An message received: ${chalk.underline(msg.id)}`));
 
@@ -262,7 +262,7 @@ export default class 藍 {
 		//#endregion
 
 		if (!immediate) {
-			await delay(1000);
+			await sleep(1000);
 		}
 
 		// リアクションする
@@ -274,7 +274,7 @@ export default class 藍 {
 		}
 	}
 
-	@autobind
+	@bindThis
 	private onNotification(notification: any) {
 		switch (notification.type) {
 			// リアクションされたら親愛度を少し上げる
@@ -290,7 +290,7 @@ export default class 藍 {
 		}
 	}
 
-	@autobind
+	@bindThis
 	private crawleTimer() {
 		const timers = this.timers.find();
 		for (const timer of timers) {
@@ -303,7 +303,7 @@ export default class 藍 {
 		}
 	}
 
-	@autobind
+	@bindThis
 	private logWaking() {
 		this.setMeta({
 			lastWakingAt: Date.now(),
@@ -313,7 +313,7 @@ export default class 藍 {
 	/**
 	 * データベースのコレクションを取得します
 	 */
-	@autobind
+	@bindThis
 	public getCollection(name: string, opts?: any): loki.Collection {
 		let collection: loki.Collection;
 
@@ -326,7 +326,7 @@ export default class 藍 {
 		return collection;
 	}
 
-	@autobind
+	@bindThis
 	public lookupFriend(userId: User['id']): Friend | null {
 		const doc = this.friends.findOne({
 			userId: userId
@@ -342,9 +342,9 @@ export default class 藍 {
 	/**
 	 * ファイルをドライブにアップロードします
 	 */
-	@autobind
+	@bindThis
 	public async upload(file: Buffer | fs.ReadStream, meta: any) {
-		const res = await request.post({
+		const res = await got.post({
 			url: `${config.apiUrl}/drive/files/create`,
 			formData: {
 				i: config.i,
@@ -354,14 +354,14 @@ export default class 藍 {
 				}
 			},
 			json: true
-		});
+		}).json();
 		return res;
 	}
 
 	/**
 	 * 投稿します
 	 */
-	@autobind
+	@bindThis
 	public async post(param: any) {
 		const res = await this.api('notes/create', param);
 		return res.createdNote;
@@ -370,23 +370,25 @@ export default class 藍 {
 	/**
 	 * 指定ユーザーにトークメッセージを送信します
 	 */
-	@autobind
+	@bindThis
 	public sendMessage(userId: any, param: any) {
-		return this.api('messaging/messages/create', Object.assign({
-			userId: userId,
+		return this.post(Object.assign({
+			visibility: 'specified',
+			visibleUserIds: [userId],
 		}, param));
 	}
 
 	/**
 	 * APIを呼び出します
 	 */
-	@autobind
+	@bindThis
 	public api(endpoint: string, param?: any) {
-		return request.post(`${config.apiUrl}/${endpoint}`, {
+		this.log(`API: ${endpoint}`);
+		return got.post(`${config.apiUrl}/${endpoint}`, {
 			json: Object.assign({
 				i: config.i
 			}, param)
-		});
+		}).json();
 	};
 
 	/**
@@ -396,7 +398,7 @@ export default class 藍 {
 	 * @param id トークメッセージ上のコンテキストならばトーク相手のID、そうでないなら待ち受ける投稿のID
 	 * @param data コンテキストに保存するオプションのデータ
 	 */
-	@autobind
+	@bindThis
 	public subscribeReply(module: Module, key: string | null, id: string, data?: any) {
 		this.contexts.insertOne({
 			noteId: id,
@@ -411,7 +413,7 @@ export default class 藍 {
 	 * @param module 解除するモジュール名
 	 * @param key コンテキストを識別するためのキー
 	 */
-	@autobind
+	@bindThis
 	public unsubscribeReply(module: Module, key: string | null) {
 		this.contexts.findAndRemove({
 			key: key,
@@ -426,7 +428,7 @@ export default class 藍 {
 	 * @param delay ミリ秒
 	 * @param data オプションのデータ
 	 */
-	@autobind
+	@bindThis
 	public setTimeoutWithPersistence(module: Module, delay: number, data?: any) {
 		const id = uuid();
 		this.timers.insertOne({
@@ -440,7 +442,7 @@ export default class 藍 {
 		this.log(`Timer persisted: ${module.name} ${id} ${delay}ms`);
 	}
 
-	@autobind
+	@bindThis
 	public getMeta() {
 		const rec = this.meta.findOne();
 
@@ -456,7 +458,7 @@ export default class 藍 {
 		}
 	}
 
-	@autobind
+	@bindThis
 	public setMeta(meta: Partial<Meta>) {
 		const rec = this.getMeta();
 
