@@ -1,5 +1,6 @@
 type Config = {
 	host: string;
+	serverName?: string;
 	i: string;
 	master?: string;
 	wsUrl: string;
@@ -9,12 +10,14 @@ type Config = {
 	notingEnabled: boolean;
 	chartEnabled: boolean;
 	serverMonitoring: boolean;
+	checkEmojisEnabled?: boolean;
+	checkEmojisAtOnce?: boolean;
 	mecab?: string;
 	mecabDic?: string;
 	memoryDir?: string;
 };
 
-const config = require('../config.json');
+import config from '../config.json' assert { type: 'json' };
 
 config.wsUrl = config.host.replace('http', 'ws');
 config.apiUrl = config.host + '/api';
