@@ -1,13 +1,13 @@
-import autobind from 'autobind-decorator';
-import * as chalk from 'chalk';
-const delay = require('timeout-as-promise');
+import { bindThis } from '@/decorators.js';
+import chalk from 'chalk';
 
-import 藍 from '@/ai';
-import Friend from '@/friend';
-import { User } from '@/misskey/user';
-import includes from '@/utils/includes';
-import or from '@/utils/or';
-import config from '@/config';
+import 藍 from '@/ai.js';
+import Friend from '@/friend.js';
+import { User } from '@/misskey/user.js';
+import includes from '@/utils/includes.js';
+import or from '@/utils/or.js';
+import config from '@/config.js';
+import { sleep } from '@/utils/sleep.js';
 
 export default class Message {
 	private ai: 藍;
@@ -68,7 +68,7 @@ export default class Message {
 		});
 	}
 
-	@autobind
+	@bindThis
 	public async reply(text: string | null, opts?: {
 		file?: any;
 		cw?: string;
@@ -80,7 +80,7 @@ export default class Message {
 		this.ai.log(`>>> Sending reply to ${chalk.underline(this.id)}`);
 
 		if (!opts?.immediate) {
-			await delay(2000);
+			await sleep(2000);
 		}
 
 		return await this.ai.post({
@@ -92,12 +92,12 @@ export default class Message {
 		});
 	}
 
-	@autobind
+	@bindThis
 	public includes(words: string[]): boolean {
 		return includes(this.text, words);
 	}
 
-	@autobind
+	@bindThis
 	public or(words: (string | RegExp)[]): boolean {
 		return or(this.text, words);
 	}
