@@ -1,8 +1,8 @@
-import autobind from 'autobind-decorator';
-import * as loki from 'lokijs';
-import Module from '@/module';
-import Message from '@/message';
-import serifs from '@/serifs';
+import { bindThis } from '@/decorators.js';
+import loki from 'lokijs';
+import Module from '@/module.js';
+import Message from '@/message.js';
+import serifs from '@/serifs.js';
 
 export default class extends Module {
 	public readonly name = 'guessingGame';
@@ -16,7 +16,7 @@ export default class extends Module {
 		endedAt: number | null;
 	}>;
 
-	@autobind
+	@bindThis
 	public install() {
 		this.guesses = this.ai.getCollection('guessingGame', {
 			indices: ['userId']
@@ -28,7 +28,7 @@ export default class extends Module {
 		};
 	}
 
-	@autobind
+	@bindThis
 	private async mentionHook(msg: Message) {
 		if (!msg.includes(['数当て', '数あて'])) return false;
 
@@ -55,7 +55,7 @@ export default class extends Module {
 		return true;
 	}
 
-	@autobind
+	@bindThis
 	private async contextHook(key: any, msg: Message) {
 		if (msg.text == null) return;
 
