@@ -1,10 +1,10 @@
-import autobind from 'autobind-decorator';
-import * as loki from 'lokijs';
-import Module from '@/module';
-import Message from '@/message';
-import serifs, { getSerif } from '@/serifs';
-import { acct } from '@/utils/acct';
-import config from '@/config';
+import { bindThis } from '@/decorators.js';
+import loki from 'lokijs';
+import Module from '@/module.js';
+import Message from '@/message.js';
+import serifs, { getSerif } from '@/serifs.js';
+import { acct } from '@/utils/acct.js';
+import config from '@/config.js';
 
 const NOTIFY_INTERVAL = 1000 * 60 * 60 * 12;
 
@@ -20,7 +20,7 @@ export default class extends Module {
 		createdAt: number;
 	}>;
 
-	@autobind
+	@bindThis
 	public install() {
 		this.reminds = this.ai.getCollection('reminds', {
 			indices: ['userId', 'id']
@@ -33,7 +33,7 @@ export default class extends Module {
 		};
 	}
 
-	@autobind
+	@bindThis
 	private async mentionHook(msg: Message) {
 		let text = msg.extractedText.toLowerCase();
 		if (!text.startsWith('remind') && !text.startsWith('todo')) return false;
@@ -98,7 +98,7 @@ export default class extends Module {
 		};
 	}
 
-	@autobind
+	@bindThis
 	private async contextHook(key: any, msg: Message, data: any) {
 		if (msg.text == null) return;
 
@@ -128,7 +128,7 @@ export default class extends Module {
 		}
 	}
 
-	@autobind
+	@bindThis
 	private async timeoutCallback(data) {
 		const remind = this.reminds.findOne({
 			id: data.id
