@@ -1,11 +1,16 @@
 import * as childProcess from 'child_process';
-import autobind from 'autobind-decorator';
-import Module from '@/module';
-import serifs from '@/serifs';
-import config from '@/config';
-import Message from '@/message';
-import Friend from '@/friend';
-import getDate from '@/utils/get-date';
+import { bindThis } from '@/decorators.js';
+import Module from '@/module.js';
+import serifs from '@/serifs.js';
+import config from '@/config.js';
+import Message from '@/message.js';
+import Friend from '@/friend.js';
+import getDate from '@/utils/get-date.js';
+import { fileURLToPath } from 'node:url';
+import { dirname, resolve } from 'node:path';
+
+const _filename = fileURLToPath(import.meta.url);
+const _dirname = dirname(_filename);
 
 export default class extends Module {
 	public readonly name = 'reversi';
@@ -15,17 +20,17 @@ export default class extends Module {
 	 */
 	private reversiConnection?: any;
 
-	@autobind
+	@bindThis
 	public install() {
 		if (!config.reversiEnabled) return {};
 
-		this.reversiConnection = this.ai.connection.useSharedConnection('gamesReversi');
+		this.reversiConnection = this.ai.connection.useSharedConnection('reversi');
 
 		// 招待されたとき
-		this.reversiConnection.on('invited', msg => this.onReversiInviteMe(msg.parent));
+		this.reversiConnection.on('invited', msg => this.onReversiInviteMe(msg.user));
 
 		// マッチしたとき
-		this.reversiConnection.on('matched', msg => this.onReversiGameStart(msg));
+		this.reversiConnection.on('matched', msg => this.onReversiGameStart(msg.game));
 
 		if (config.reversiEnabled) {
 			const mainStream = this.ai.connection.useSharedConnection('main');
@@ -43,13 +48,17 @@ export default class extends Module {
 		};
 	}
 
-	@autobind
+	@bindThis
 	private async mentionHook(msg: Message) {
 		if (msg.includes(['リバーシ', 'オセロ', 'reversi', 'othello'])) {
 			if (config.reversiEnabled) {
 				msg.reply(serifs.reversi.ok);
 
-				this.ai.api('games/reversi/match', {
+				if (msg.includes(['接待'])) {
+					msg.friend.updateReversiStrength(0);
+				}
+
+				this.ai.api('reversi/match', {
 					userId: msg.userId
 				});
 			} else {
@@ -62,13 +71,13 @@ export default class extends Module {
 		}
 	}
 
-	@autobind
+	@bindThis
 	private async onReversiInviteMe(inviter: any) {
 		this.log(`Someone invited me: @${inviter.username}`);
 
 		if (config.reversiEnabled) {
 			// 承認
-			const game = await this.ai.api('games/reversi/match', {
+			const game = await this.ai.api('reversi/match', {
 				userId: inviter.id
 			});
 
@@ -78,12 +87,19 @@ export default class extends Module {
 		}
 	}
 
-	@autobind
+	@bindThis
 	private onReversiGameStart(game: any) {
-		this.log('enter reversi game room');
+		let strength = 5;
+		const friend = this.ai.lookupFriend(game.user1Id !== this.ai.account.id ? game.user1Id : game.user2Id)!;
+		if (friend != null) {
+			strength = friend.doc.reversiStrength ?? 5;
+			friend.updateReversiStrength(null);
+		}
+
+		this.log(`enter reversi game room: ${game.id}`);
 
 		// ゲームストリームに接続
-		const gw = this.ai.connection.connectToChannel('gamesReversiGame', {
+		const gw = this.ai.connection.connectToChannel('reversiGame', {
 			gameId: game.id
 		});
 
@@ -92,12 +108,12 @@ export default class extends Module {
 			id: 'publish',
 			type: 'switch',
 			label: '藍が対局情報を投稿するのを許可',
-			value: true
+			value: true,
 		}, {
 			id: 'strength',
 			type: 'radio',
 			label: '強さ',
-			value: 3,
+			value: strength,
 			items: [{
 				label: '接待',
 				value: 0
@@ -117,7 +133,7 @@ export default class extends Module {
 		}];
 
 		//#region バックエンドプロセス開始
-		const ai = childProcess.fork(__dirname + '/back.js');
+		const ai = childProcess.fork(_dirname + '/back.js');
 
 		// バックエンドプロセスに情報を渡す
 		ai.send({
@@ -130,9 +146,10 @@ export default class extends Module {
 		});
 
 		ai.on('message', (msg: Record<string, any>) => {
-			if (msg.type == 'put') {
-				gw.send('set', {
-					pos: msg.pos
+			if (msg.type == 'putStone') {
+				gw.send('putStone', {
+					pos: msg.pos,
+					id: msg.id,
 				});
 			} else if (msg.type == 'ended') {
 				gw.dispose();
@@ -144,21 +161,26 @@ export default class extends Module {
 		// ゲームストリームから情報が流れてきたらそのままバックエンドプロセスに伝える
 		gw.addListener('*', message => {
 			ai.send(message);
+
+			if (message.type === 'updateSettings') {
+				if (message.body.key === 'canPutEverywhere') {
+					if (message.body.value === true) {
+						gw.send('ready', false);
+					} else {
+						gw.send('ready', true);
+					}
+				}
+			}
 		});
 		//#endregion
 
-		// フォーム初期化
-		setTimeout(() => {
-			gw.send('initForm', form);
-		}, 1000);
-
 		// どんな設定内容の対局でも受け入れる
 		setTimeout(() => {
-			gw.send('accept', {});
-		}, 2000);
+			gw.send('ready', true);
+		}, 1000);
 	}
 
-	@autobind
+	@bindThis
 	private onGameEnded(game: any) {
 		const user = game.user1Id == this.ai.account.id ? game.user2 : game.user1;
 
