@@ -1,21 +1,21 @@
-import autobind from 'autobind-decorator';
-import { HandlerResult } from '@/ai';
-import Module from '@/module';
-import Message from '@/message';
-import serifs, { getSerif } from '@/serifs';
-import getDate from '@/utils/get-date';
+import { bindThis } from '@/decorators.js';
+import { HandlerResult } from '@/ai.js';
+import Module from '@/module.js';
+import Message from '@/message.js';
+import serifs, { getSerif } from '@/serifs.js';
+import getDate from '@/utils/get-date.js';
 
 export default class extends Module {
 	public readonly name = 'talk';
 
-	@autobind
+	@bindThis
 	public install() {
 		return {
 			mentionHook: this.mentionHook,
 		};
 	}
 
-	@autobind
+	@bindThis
 	private async mentionHook(msg: Message) {
 		if (!msg.text) return false;
 
@@ -37,7 +37,7 @@ export default class extends Module {
 		);
 	}
 
-	@autobind
+	@bindThis
 	private greet(msg: Message): boolean {
 		if (msg.text == null) return false;
 
@@ -106,7 +106,7 @@ export default class extends Module {
 		return false;
 	}
 
-	@autobind
+	@bindThis
 	private erait(msg: Message): boolean {
 		const match = msg.extractedText.match(/(.+?)た(から|ので)(褒|ほ)めて/);
 		if (match) {
@@ -133,7 +133,7 @@ export default class extends Module {
 		return true;
 	}
 
-	@autobind
+	@bindThis
 	private omedeto(msg: Message): boolean {
 		if (!msg.includes(['おめでと'])) return false;
 
@@ -142,7 +142,7 @@ export default class extends Module {
 		return true;
 	}
 
-	@autobind
+	@bindThis
 	private nadenade(msg: Message): boolean {
 		if (!msg.includes(['なでなで'])) return false;
 
@@ -174,7 +174,7 @@ export default class extends Module {
 		return true;
 	}
 
-	@autobind
+	@bindThis
 	private kawaii(msg: Message): boolean {
 		if (!msg.includes(['かわいい', '可愛い'])) return false;
 
@@ -186,7 +186,7 @@ export default class extends Module {
 		return true;
 	}
 
-	@autobind
+	@bindThis
 	private suki(msg: Message): boolean {
 		if (!msg.or(['好き', 'すき'])) return false;
 
@@ -198,7 +198,7 @@ export default class extends Module {
 		return true;
 	}
 
-	@autobind
+	@bindThis
 	private hug(msg: Message): boolean {
 		if (!msg.or(['ぎゅ', 'むぎゅ', /^はぐ(し(て|よ|よう)?)?$/])) return false;
 
@@ -229,7 +229,7 @@ export default class extends Module {
 		return true;
 	}
 
-	@autobind
+	@bindThis
 	private humu(msg: Message): boolean {
 		if (!msg.includes(['踏んで'])) return false;
 
@@ -241,7 +241,7 @@ export default class extends Module {
 		return true;
 	}
 
-	@autobind
+	@bindThis
 	private batou(msg: Message): boolean {
 		if (!msg.includes(['罵倒して', '罵って'])) return false;
 
@@ -253,7 +253,7 @@ export default class extends Module {
 		return true;
 	}
 
-	@autobind
+	@bindThis
 	private itai(msg: Message): boolean {
 		if (!msg.or(['痛い', 'いたい']) && !msg.extractedText.endsWith('痛い')) return false;
 
@@ -262,7 +262,7 @@ export default class extends Module {
 		return true;
 	}
 
-	@autobind
+	@bindThis
 	private ote(msg: Message): boolean {
 		if (!msg.or(['お手'])) return false;
 
@@ -274,7 +274,7 @@ export default class extends Module {
 		return true;
 	}
 
-	@autobind
+	@bindThis
 	private ponkotu(msg: Message): boolean | HandlerResult {
 		if (!msg.includes(['ぽんこつ'])) return false;
 
@@ -285,7 +285,7 @@ export default class extends Module {
 		};
 	}
 
-	@autobind
+	@bindThis
 	private rmrf(msg: Message): boolean | HandlerResult {
 		if (!msg.includes(['rm -rf'])) return false;
 
@@ -296,7 +296,7 @@ export default class extends Module {
 		};
 	}
 
-	@autobind
+	@bindThis
 	private shutdown(msg: Message): boolean | HandlerResult {
 		if (!msg.includes(['shutdown'])) return false;
 
