@@ -1,4 +1,4 @@
-import { katakanaToHiragana, hankakuToZenkaku } from './japanese';
+import { katakanaToHiragana, hankakuToZenkaku } from './japanese.js';
 
 export default function(text: string, words: string[]): boolean {
 	if (text == null) return false;
