@@ -1,4 +1,4 @@
-import { hankakuToZenkaku, katakanaToHiragana } from './japanese';
+import { hankakuToZenkaku, katakanaToHiragana } from './japanese.js';
 
 export default function(text: string, words: (string | RegExp)[]): boolean {
 	if (text == null) return false;
